again a 7 kyu that's tough middleElement 11/1/2021
Mega summer daily codewars. After the stand-up, dove deep into the thought process for the solution. Why the sort method was used, why a-b, why spread syntax instead of keeping it as an array, why there is bracket notation with [1], what indexOf does. Thanks for coming to my Ted Talk.